// src/types/response-codes.ts
export enum Success {
  OK = 200,
  CREATED = 201,
  NO_CONTENT = 204,
}

export enum ClientError {
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
}

export enum ServerError {
  INTERNAL_SERVER_ERROR = 500,
  NOT_IMPLEMENTED = 501,
  SERVICE_UNAVAILABLE = 503,
}